Split protocol logic from core into tsrp
