--- zodiac-core/src/Zodiac/Core.hs
@@ -1,9 +1,0 @@
-{-# LANGUAGE NoImplicitPrelude #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE OverloadedStrings #-}
-module Zodiac.Core(
-    module X
-  ) where
-
-import           Zodiac.Core.Data.Key as X
-import           Zodiac.Core.Key as X
