+++ zodiac-tsrp/ambiata-zodiac-tsrp.cabal
@@ -1,0 +1,138 @@
+name:                  ambiata-zodiac-tsrp
+version:               0.0.1
+license:               Apache-2.0
+author:                Ambiata <info@ambiata.com>
+maintainer:            Ambiata <info@ambiata.com>
+copyright:             (c) 2016 Ambiata.
+synopsis:              ambiata-zodiac-tsrp
+category:              System
+cabal-version:         >= 1.8
+build-type:            Simple
+description:           ambiata-zodiac-tsrp
+
+library
+  build-depends:
+                       base                            >= 3          && < 5
+                     , ambiata-p
+                     , ambiata-tinfoil
+                     , ambiata-x-eithert
+                     , ambiata-zodiac-core
+                     , attoparsec                      >= 0.10 && < 0.13
+                     , base16-bytestring               == 0.1.1.*
+                     , bytestring                      == 0.10.*
+                     , deepseq-generics                == 0.2.0.*
+                     , http-types                      >= 0.8        && < 0.10
+                     , semigroups                      >= 0.16       && < 0.19
+                     , text                            == 1.2.*
+                     , time                            >= 1.5        && < 1.7
+                     , transformers                    >= 0.4        && < 0.6
+
+  ghc-options:
+                       -Wall
+
+  hs-source-dirs:
+                       src
+
+  exposed-modules:
+                       Zodiac.TSRP.Data
+                       Zodiac.TSRP.Data.Error
+                       Zodiac.TSRP.Data.Key
+                       Zodiac.TSRP.Data.Symmetric
+                       Zodiac.TSRP.Key
+                       Zodiac.TSRP.MAC
+                       Zodiac.TSRP.Symmetric
+
+test-suite test
+  type:                exitcode-stdio-1.0
+
+  main-is:             test.hs
+
+  ghc-options:         -Wall -threaded -O2
+
+  hs-source-dirs:
+                       test
+
+  build-depends:
+                       base                            >= 3          && < 5
+                     , QuickCheck                      >= 2.8.2      && < 2.9
+                     , ambiata-disorder-core
+                     , ambiata-disorder-corpus
+                     , ambiata-p
+                     , ambiata-tinfoil
+                     , ambiata-tinfoil-test
+                     , ambiata-zodiac-core
+                     , ambiata-zodiac-core-test
+                     , ambiata-zodiac-tsrp
+                     , bytestring                      == 0.10.*
+                     , containers                      == 0.5.*
+                     , http-types                      >= 0.8        && < 0.10
+                     , quickcheck-instances            == 0.3.*
+                     , random                          == 1.1.*
+                     , semigroups                      >= 0.16       && < 0.19
+                     , text                            == 1.2.*
+                     , time                            >= 1.5        && < 1.7
+                     , transformers                    >= 0.4        && < 0.6
+
+test-suite test-io
+  type:                exitcode-stdio-1.0
+
+  main-is:             test-io.hs
+
+  ghc-options:         -Wall -threaded -O2
+
+  hs-source-dirs:
+                       test
+
+  build-depends:
+                       base                            >= 3          && < 5
+                     , QuickCheck                      >= 2.8.2      && < 2.9
+                     , ambiata-disorder-core
+                     , ambiata-disorder-corpus
+                     , ambiata-p
+                     , ambiata-tinfoil
+                     , ambiata-tinfoil-test
+                     , ambiata-zodiac-core
+                     , ambiata-zodiac-core-test
+                     , ambiata-zodiac-tsrp
+                     , bytestring                      == 0.10.*
+                     , containers                      == 0.5.*
+                     , http-types                      >= 0.8        && < 0.10
+                     , quickcheck-instances            == 0.3.*
+                     , random                          == 1.1.*
+                     , semigroups                      >= 0.16       && < 0.19
+                     , text                            == 1.2.*
+                     , time                            >= 1.5        && < 1.7
+                     , transformers                    >= 0.4        && < 0.6
+
+benchmark bench
+  type:                exitcode-stdio-1.0
+
+  main-is:             bench.hs
+
+  ghc-options:         -Wall -threaded -O2
+
+  hs-source-dirs:
+                       bench
+                     , test
+
+  build-depends:
+                       base                            >= 3          && < 5
+                     , QuickCheck                      >= 2.8.2      && < 2.9
+                     , ambiata-disorder-core
+                     , ambiata-disorder-corpus
+                     , ambiata-p
+                     , ambiata-tinfoil
+                     , ambiata-tinfoil-test
+                     , ambiata-zodiac-core
+                     , ambiata-zodiac-core-test
+                     , ambiata-zodiac-tsrp
+                     , bytestring                      == 0.10.*
+                     , containers                      == 0.5.*
+                     , criterion                       == 1.1.*
+                     , http-types                      >= 0.8        && < 0.10
+                     , quickcheck-instances            == 0.3.*
+                     , random                          == 1.1.*
+                     , semigroups                      >= 0.16       && < 0.19
+                     , text                            == 1.2.*
+                     , time                            >= 1.5        && < 1.7
+                     , transformers                    >= 0.4        && < 0.6
