+++ zodiac-tsrp/src/Zodiac/TSRP/Data.hs
@@ -1,8 +1,11 @@
 {-# LANGUAGE NoImplicitPrelude #-}
-module Zodiac.Core.Data(
+module Zodiac.TSRP.Data(
       module X
   ) where
 
+import           Zodiac.TSRP.Data.Error as X
+import           Zodiac.TSRP.Data.Key as X
 import           Zodiac.Core.Data.Protocol as X
 import           Zodiac.Core.Data.Request as X
+import           Zodiac.TSRP.Data.Symmetric as X
 import           Zodiac.Core.Data.Time as X
