+++ zodiac-tsrp/src/Zodiac/TSRP/MAC.hs
@@ -6,7 +6,7 @@ HMAC functionality for symmetric request-signing.
 
 Don't use this module directly unless you know what you're doing.
 -}
-module Zodiac.Core.MAC(
+module Zodiac.TSRP.MAC(
     deriveRequestKey
   ) where
 
@@ -15,7 +15,7 @@ import           P
 import           Tinfoil.Data (SymmetricKey(..), MAC(..))
 import           Tinfoil.MAC (hmacSHA256)
 
-import           Zodiac.Core.Data
+import           Zodiac.TSRP.Data
 
 -- | Derive the key we actually use to authenticate the request from the
 -- secret key, the key ID and the date part of the request timestamp
