+++ zodiac-tsrp/test/Test/IO/Zodiac/TSRP/Symmetric.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell #-}
-module Test.IO.Zodiac.Core.Symmetric where
+module Test.IO.Zodiac.TSRP.Symmetric where
 
 import           Data.Time.Clock (addUTCTime)
 
@@ -15,14 +15,14 @@ import           System.IO (IO)
 
 import           Test.QuickCheck
 import           Test.QuickCheck.Instances ()
-import           Test.Zodiac.Core.Arbitrary ()
+import           Test.Zodiac.TSRP.Arbitrary ()
 import           Test.Zodiac.Core.Gen
 
 import           Tinfoil.Data
 
-import           Zodiac.Core.Data
+import           Zodiac.TSRP.Data
 import           Zodiac.Core.Request
-import           Zodiac.Core.Symmetric
+import           Zodiac.TSRP.Symmetric
 
 prop_verifyRequest' :: SymmetricProtocol
                     -> KeyId
