+++ zodiac-tsrp/test/test.hs
@@ -1,0 +1,9 @@
+import           Disorder.Core.Main
+
+import qualified Test.Zodiac.TSRP.MAC
+
+main :: IO ()
+main =
+  disorderMain [
+    Test.Zodiac.TSRP.MAC.tests
+  ]
